Refactor jinja into 2 crates, one for parserDB types and one for runtime/prompt-rendering concepts
