+++ engine/baml-lib/baml-core/Cargo.toml
@@ -16,7 +16,7 @@ enumflags2 = "0.7"
 log = "0.4.20"
 indexmap.workspace = true
 internal-baml-diagnostics = { path = "../diagnostics" }
-internal-baml-jinja.workspace = true
+internal-baml-jinja-types = { path = "../jinja" }
 internal-baml-parser-database = { path = "../parser-database" }
 internal-baml-prompt-parser = { path = "../prompt-parser" }
 internal-baml-schema-ast = { path = "../schema-ast" }
