+++ engine/baml-lib/baml-core/src/validate/validation_pipeline/validations/classes.rs
@@ -5,7 +5,7 @@ use crate::validate::validation_pipeline::context::Context;
 use internal_baml_diagnostics::DatamodelError;
 
 pub(super) fn validate(ctx: &mut Context<'_>) {
-    let mut defined_types = internal_baml_jinja::PredefinedTypes::default();
+    let mut defined_types = internal_baml_jinja_types::PredefinedTypes::default();
 
     for cls in ctx.db.walk_classes() {
         for c in cls.static_fields() {
