+++ engine/baml-lib/baml-types/Cargo.toml
@@ -7,6 +7,7 @@ description.workspace = true
 license-file.workspace = true
 
 [dependencies]
+# internal-baml-core = { path = "../baml-core" }
 anyhow.workspace = true
 clap.workspace = true
 derive_builder.workspace = true
