+++ engine/baml-lib/jsonish/Cargo.toml
@@ -8,7 +8,8 @@ description.workspace = true
 license-file.workspace = true
 
 [dependencies]
-internal-baml-jinja = { path = "../jinja" }
+internal-baml-jinja = { path = "../jinja-runtime" }
+# internal-baml-jinja-types = { path = "../jinja" }
 internal-baml-core = { path = "../baml-core" }
 baml-types = { path = "../baml-types" }
 colored = "2"
