+++ engine/baml-lib/parser-database/src/walkers/template_string.rs
@@ -1,5 +1,5 @@
 use either::Either;
-use internal_baml_jinja::{PredefinedTypes, Type};
+use internal_baml_jinja_types::{PredefinedTypes, Type};
 use internal_baml_schema_ast::ast::{self, BlockArgs, Span, WithIdentifier, WithName, WithSpan};
 
 use crate::types::TemplateStringProperties;
