+++ engine/baml-runtime/src/internal/prompt_renderer/mod.rs
@@ -94,6 +94,7 @@ impl PromptRenderer {
                     template: t.template().into(),
                 })
                 .collect::<Vec<_>>(),
+            ir,
         )
     }
 }
